langchain/requirements.txt
langchain
ipykernel
langchain-openai
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
langchain-text-splitters
lxml
openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
streamlit
langchain-ollama
langchain_groq
fastapi
gunicorn
langserve
uvicorn
sse_starlette
unstructured
unstructured[pdf]
wikipedia
duckduckgo-search